Limpieza de clases no utilizadas, mejora del esquema de Base de Datos paso 1

+++ src/main/java/com/es/unex/cum/mdai/Steem/Modelo/Usuario.java
@@ -13,7 +13,7 @@ public abstract class Usuario {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected int idUsuario;
+    protected long idUsuario;
 
     @Column(nullable = false, unique = true)
     protected String nombreUsuario;
@@ -21,7 +21,7 @@ public abstract class Usuario {
     protected String password;
     protected Date fechaRegistro;
 
-    public int getIdUsuario() {
+    public long getIdUsuario() {
         return idUsuario;
     }
